refactor: extract `AssetList`
